requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-admin==2.0.2
django-excel-response2==3.0.6
django-rest-swagger==2.2.0
django-six==1.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
excel-base==1.0.4
idna==3.10
inflection==0.5.1
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
screen==1.0.1
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
TimeConvert==3.0.13
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
xlwt==1.3.0
